// src/app/api/model/classification.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * A classification system that groups related objects together.
 */
export const Classification = {
    Object: 'object',
    Device: 'device',
    Integration: 'integration',
    Controller: 'controller',
    Point: 'point',
    Site: 'site',
    NavList: 'navList',
    Extension: 'extension',
    Folder: 'folder',
    Reference: 'reference',
    Server: 'server',
    Archive: 'archive'
} as const;
export type Classification = typeof Classification[keyof typeof Classification];