// src/app/api/model/getObjectAttributeResponse.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConditionValue } from './conditionValue';


export interface GetObjectAttributeResponse { 
    /**
     * The `item` section of the response contains the attributes of the object. In this case it will contain only the specified attribute. 
     */
    item: object;
    /**
     * If included in the response, this is the JSON schema that describes the properties of the `item` property 
     */
    readonly schema?: object;
    /**
     * This property contains condition flags (if any are set) for the properties of `item`. If none of the properties have any condition flags set, then `condition` will not be included in the response or it will be empty. 
     */
    readonly condition: { [key: string]: ConditionValue; };
}