// src/app/api/model/networkDevices.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NetworkDevicesEffectivePermissions } from './networkDevicesEffectivePermissions';
import { NetworkDevicesItemsInner } from './networkDevicesItemsInner';


export interface NetworkDevices { 
    /**
     * The total number of items in the collection.
     */
    total: number;
    /**
     * The collection of network devices. Incomplete network devices are only included if the `includeIncomplete` query parameter is set to `true` and the API is at release 15.0 or greater. 
     */
    items: Array<NetworkDevicesItemsInner>;
    /**
     * A link to the next page of results.
     */
    next: string | null;
    /**
     * A link to the previous page of results.
     */
    previous: string | null;
    /**
     * A link to this equipment collection.
     */
    self: string;
    effectivePermissions: NetworkDevicesEffectivePermissions;
}