// src/app/api/model/postObjectsBatch200Response.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PostObjectsBatch200ResponseResponsesInner } from './postObjectsBatch200ResponseResponsesInner';


export interface PostObjectsBatch200Response { 
    responses: Array<PostObjectsBatch200ResponseResponsesInner>;
}