// src/app/api/model/postObjectsBatch200ResponseResponsesInner.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PostObjectsBatch200ResponseResponsesInnerBody } from './postObjectsBatch200ResponseResponsesInnerBody';


export interface PostObjectsBatch200ResponseResponsesInner { 
    /**
     * One of the client supplied unique request identifiers in the request. This identifier is used to match the response to the correct operation in the request.
     */
    id: string;
    /**
     * The HTTP Status code returned for the corresponding operation.
     */
    status: number;
    body: PostObjectsBatch200ResponseResponsesInnerBody | null;
}