// src/app/api/model/postObjectsBatchRequest.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PostObjectsBatchRequestRequestsInner } from './postObjectsBatchRequestRequestsInner';


export interface PostObjectsBatchRequest { 
    /**
     * The HTTP method to use for each request. This property is restricted to `GET` for this version of the API since _Get attribute value_ is the only operation supported in a batch.  
     */
    method: PostObjectsBatchRequest.MethodEnum;
    /**
     * A collection of entries that represent the operations to invoke.
     */
    requests: Array<PostObjectsBatchRequestRequestsInner>;
}
export namespace PostObjectsBatchRequest {
    export const MethodEnum = {
        Get: 'GET',
        Patch: 'PATCH',
        Put: 'PUT'
    } as const;
    export type MethodEnum = typeof MethodEnum[keyof typeof MethodEnum];
}